agents/search-competitors/requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools-browser
